+++ src/GenJet.C
@@ -12,9 +12,9 @@ GenJet::GenJet(Collection& c, unsigned short i, short j ):
 
 // Kinematic variables
 
-double & GenJet::Pt  () { return m_collection -> GetData() -> GenJetPtAK4  -> at ( m_raw_index ); }
-double & GenJet::Eta () { return m_collection -> GetData() -> GenJetEtaAK4 -> at ( m_raw_index ); } 
-double & GenJet::Phi () { return m_collection -> GetData() -> GenJetPhiAK4 -> at ( m_raw_index ); } 
+float & GenJet::Pt  () { return m_collection -> GetData() -> GenJetPtAK4  -> at ( m_raw_index ); }
+float & GenJet::Eta () { return m_collection -> GetData() -> GenJetEtaAK4 -> at ( m_raw_index ); } 
+float & GenJet::Phi () { return m_collection -> GetData() -> GenJetPhiAK4 -> at ( m_raw_index ); } 
 
 std::ostream& operator<<(std::ostream& stream, GenJet& object) {
   stream << object.Name() << " " << ": "
