+++ src/PFJet.C
@@ -11,27 +11,27 @@ PFJet::PFJet (Collection& c, unsigned int i, short j ):
   Object ( c,i, "PFJet" )
 {}
                                      
-double & PFJet::Pt                       () { return m_collection -> GetData() -> PFJetPtAK4CHS                           -> at ( m_raw_index ); } 
-double & PFJet::Eta                      () { return m_collection -> GetData() -> PFJetEtaAK4CHS                          -> at ( m_raw_index ); } 
-double & PFJet::Phi                      () { return m_collection -> GetData() -> PFJetPhiAK4CHS                          -> at ( m_raw_index ); } 
-double PFJet::Energy                     () { return m_collection -> GetData() -> PFJetEnergyAK4CHS                       -> at ( m_raw_index ); }
-double PFJet::JECUnc		         () { return m_collection -> GetData() -> PFJetJECUncAK4CHS                       -> at ( m_raw_index ); }
+float & PFJet::Pt                       () { return m_collection -> GetData() -> PFJetPtAK4CHS                           -> at ( m_raw_index ); } 
+float & PFJet::Eta                      () { return m_collection -> GetData() -> PFJetEtaAK4CHS                          -> at ( m_raw_index ); } 
+float & PFJet::Phi                      () { return m_collection -> GetData() -> PFJetPhiAK4CHS                          -> at ( m_raw_index ); } 
+float PFJet::Energy                     () { return m_collection -> GetData() -> PFJetEnergyAK4CHS                       -> at ( m_raw_index ); }
+float PFJet::JECUnc		         () { return m_collection -> GetData() -> PFJetJECUncAK4CHS                       -> at ( m_raw_index ); }
                                      
-double PFJet::NeutralHadronEnergyFraction() { return m_collection -> GetData() -> PFJetNeutralHadronEnergyFractionAK4CHS  -> at ( m_raw_index ); } 
-double PFJet::NeutralEmEnergyFraction    () { return m_collection -> GetData() -> PFJetNeutralEmEnergyFractionAK4CHS      -> at ( m_raw_index ); } 
+float PFJet::NeutralHadronEnergyFraction() { return m_collection -> GetData() -> PFJetNeutralHadronEnergyFractionAK4CHS  -> at ( m_raw_index ); } 
+float PFJet::NeutralEmEnergyFraction    () { return m_collection -> GetData() -> PFJetNeutralEmEnergyFractionAK4CHS      -> at ( m_raw_index ); } 
 int    PFJet::NeutralHadronMultiplicity  () { return m_collection -> GetData() -> PFJetNeutralHadronMultiplicityAK4CHS    -> at ( m_raw_index ); }
 int    PFJet::NeutralMultiplicity        () { return m_collection -> GetData() -> PFJetNeutralMultiplicityAK4CHS          -> at ( m_raw_index ); }
 int    PFJet::NConstituents              () { return m_collection -> GetData() -> PFJetNConstituentsAK4CHS                -> at ( m_raw_index ); } 
-double PFJet::ChargedHadronEnergyFraction() { return m_collection -> GetData() -> PFJetChargedHadronEnergyFractionAK4CHS  -> at ( m_raw_index ); } 
+float PFJet::ChargedHadronEnergyFraction() { return m_collection -> GetData() -> PFJetChargedHadronEnergyFractionAK4CHS  -> at ( m_raw_index ); } 
 int    PFJet::ChargedMultiplicity        () { return m_collection -> GetData() -> PFJetChargedMultiplicityAK4CHS          -> at ( m_raw_index ); } 
-double PFJet::ChargedEmEnergyFraction    () { return m_collection -> GetData() -> PFJetChargedEmEnergyFractionAK4CHS      -> at ( m_raw_index ); } 
+float PFJet::ChargedEmEnergyFraction    () { return m_collection -> GetData() -> PFJetChargedEmEnergyFractionAK4CHS      -> at ( m_raw_index ); } 
 
-double PFJet::CombinedSecondaryVertexBTag() { return m_collection -> GetData() -> PFJetCombinedSecondaryVertexBTagAK4CHS  -> at ( m_raw_index ); }
+float PFJet::CombinedSecondaryVertexBTag() { return m_collection -> GetData() -> PFJetCombinedSecondaryVertexBTagAK4CHS  -> at ( m_raw_index ); }
 
 // Energy resolution scale factors
 // see: https://twiki.cern.ch/twiki/bin/viewauth/CMS/JetResolution
-double PFJet::EnergyResScaleFactor (){ 
-  double fabs_eta = fabs ( Eta () );
+float PFJet::EnergyResScaleFactor (){ 
+  float fabs_eta = fabs ( Eta () );
   if      ( fabs_eta > 0.0 && fabs_eta <= 0.5 ) return 1.109;
   else if ( fabs_eta > 0.5 && fabs_eta <= 0.8 ) return 1.138;
   else if ( fabs_eta > 0.8 && fabs_eta <= 1.1 ) return 1.114;
@@ -48,8 +48,8 @@ double PFJet::EnergyResScaleFactor (){
 }
 
 // see: https://twiki.cern.ch/twiki/bin/viewauth/CMS/JetResolution
-double PFJet::EnergyResScaleError  (){ 
-  double fabs_eta = fabs ( Eta () );
+float PFJet::EnergyResScaleError  (){ 
+  float fabs_eta = fabs ( Eta () );
   if      ( fabs_eta > 0.0 && fabs_eta <= 0.5 ) return 0.008;
   else if ( fabs_eta > 0.5 && fabs_eta <= 0.8 ) return 0.013;
   else if ( fabs_eta > 0.8 && fabs_eta <= 1.1 ) return 0.013;
@@ -66,7 +66,7 @@ double PFJet::EnergyResScaleError  (){
 }
 
 // JES uncertainties already in the ntuple
-double PFJet::EnergyScaleFactor (){ 
+float PFJet::EnergyScaleFactor (){ 
   return JECUnc();
 }
